+++ gsonpath-compiler/src/test/java/gsonpath/generator/adapter/auto/DelimiterTest.kt
@@ -8,6 +8,7 @@ class DelimiterTest : BaseGeneratorTest() {
     @Test
     fun testStandardDelimiter() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/delimiter/standard")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestStandardDelimiter.java")
                 .addRelativeGenerated("TestStandardDelimiter_GsonTypeAdapter.java"))
     }
@@ -15,6 +16,7 @@ class DelimiterTest : BaseGeneratorTest() {
     @Test
     fun testCustomDelimiter() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/delimiter/custom")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestCustomDelimiter.java")
                 .addRelativeGenerated("TestCustomDelimiter_GsonTypeAdapter.java"))
     }
@@ -22,6 +24,7 @@ class DelimiterTest : BaseGeneratorTest() {
     @Test
     fun testMultipleDelimiters() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/delimiter/multiple")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestMultipleDelimiters.java")
                 .addRelativeGenerated("TestMultipleDelimiters_GsonTypeAdapter.java"))
     }