+++ gsonpath-compiler/src/test/java/gsonpath/generator/adapter/auto/EmptyAdapterTest.kt
@@ -18,6 +18,7 @@ class EmptyAdapterTest : BaseGeneratorTest() {
     @Test
     fun testAutoGsonAdapterOnly() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/empty/annotation_only")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestAnnotationOnly.java")
                 .addRelativeGenerated("TestAnnotationOnly_GsonTypeAdapter.java"))
     }
@@ -25,6 +26,7 @@ class EmptyAdapterTest : BaseGeneratorTest() {
     @Test
     fun testRequiresAnnotation() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/empty/ignored_fields")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestIgnoredFields.java")
                 .addRelativeGenerated("TestIgnoredFields_GsonTypeAdapter.java"))
     }
@@ -32,6 +34,7 @@ class EmptyAdapterTest : BaseGeneratorTest() {
     @Test
     fun testIgnoreInvalidFields() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/empty/invalid_fields")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestInvalidFields.java")
                 .addRelativeGenerated("TestInvalidFields_GsonTypeAdapter.java"))
     }