+++ gsonpath-compiler/src/test/java/gsonpath/generator/adapter/auto/FieldAnnotationsTest.kt
@@ -8,6 +8,7 @@ class FieldAnnotationsTest : BaseGeneratorTest() {
     @Test
     fun testFlattenJson() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/field_annotations/flatten_json")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestFlattenJson.java")
                 .addRelativeGenerated("TestFlattenJson_GsonTypeAdapter.java"))
     }
@@ -15,6 +16,7 @@ class FieldAnnotationsTest : BaseGeneratorTest() {
     @Test
     fun testExcludeFields() {
         assertGeneratedContent(BaseGeneratorTest.TestCriteria("adapter/auto/field_annotations/exclude")
+                .addAbsoluteSource("adapter/auto/TestGsonTypeFactory.java")
                 .addRelativeSource("TestExclude.java")
                 .addRelativeGenerated("TestExclude_GsonTypeAdapter.java"))
     }