+++ gsonpath-compiler/src/test/resources/adapter/loader/PackagePrivateTypeAdapterLoader.java
@@ -2,11 +2,12 @@ package adapter.loader;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
+import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import gsonpath.internal.TypeAdapterLoader;
+
 import java.lang.Override;
 
-public final class PackagePrivateTypeAdapterLoader implements TypeAdapterLoader {
+public final class PackagePrivateTypeAdapterLoader implements TypeAdapterFactory {
     @Override
     public TypeAdapter create(Gson gson, TypeToken type) {
         Class rawType = type.getRawType();