+++ gsonpath-compiler/src/test/resources/adapter/loader/TestGsonTypeFactory.java
@@ -1,0 +1,9 @@
+package adapter.loader;
+
+import com.google.gson.TypeAdapterFactory;
+
+import gsonpath.AutoGsonAdapterFactory;
+
+@AutoGsonAdapterFactory
+public interface TestGsonTypeFactory extends TypeAdapterFactory {
+}