+++ gsonpath-compiler/src/test/resources/adapter/loader/TestGsonTypeFactoryImpl.java
@@ -1,17 +1,17 @@
-package gsonpath;
+package adapter.loader;
 
 import com.google.gson.Gson;
 import com.google.gson.TypeAdapter;
+import com.google.gson.TypeAdapterFactory;
 import com.google.gson.reflect.TypeToken;
-import gsonpath.internal.TypeAdapterLoader;
 
 import java.lang.Override;
 
-public final class GeneratedTypeAdapterLoader implements TypeAdapterLoader {
-    private final TypeAdapterLoader[] mPackagePrivateLoaders;
+public final class TestGsonTypeFactoryImpl implements TestGsonTypeFactory {
+    private final TypeAdapterFactory[] mPackagePrivateLoaders;
 
-    public GeneratedTypeAdapterLoader() {
-        mPackagePrivateLoaders = new TypeAdapterLoader[3];
+    public GsonTypeAdapterLoader() {
+        mPackagePrivateLoaders = new TypeAdapterFactory[3];
         mPackagePrivateLoaders[0] = new adapter.loader.PackagePrivateTypeAdapterLoader();
         mPackagePrivateLoaders[1] = new adapter.loader.source3.PackagePrivateTypeAdapterLoader();
         mPackagePrivateLoaders[2] = new adapter.loader.source2.PackagePrivateTypeAdapterLoader();